+++ SimpleEventSourcing.Samples.Orders/Configuration.cs
@@ -1,13 +1,13 @@
 ﻿using Microsoft.Data.Sqlite;
 using System.Data;
 
-namespace SimpleEventSourcing.Samples.Orders
+namespace SimpleEventSourcing.Samples.Delivery
 {
     public static class Configuration
     {
         public static SqliteConnection GetDatabaseConnection()
         {
-            var connection = new SqliteConnection("Data Source=OrdersDatabase.db");
+            var connection = new SqliteConnection("Data Source=DeliveryDatabase.db");
             if (connection.State != ConnectionState.Open)
                 connection.Open();
