+++ SimpleEventSourcing.Samples.Orders/Properties/launchSettings.json
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "SimpleEventSourcing.Samples.Orders": {
+    "SimpleEventSourcing.Samples.Delivery": {
       "commandName": "Project",
       "launchBrowser": true,
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
