+++ SimpleEventSourcing.Samples.Orders/SimpleEventSourcing.Samples.Delivery.csproj
